+++ js/loader.js
@@ -1,6 +1,6 @@
 /*global loadFiles dashtype */
 
-var _DASHTICZ_VERSION = 92;
+var _DASHTICZ_VERSION = 93;
 var head = document.getElementsByTagName('head')[0],
     script = document.createElement('script');